Directly connect health check to GazetteerCache and remove thread
Under load from facility matching API requests our app service tasks are being
marked as unhealthy and being restarted. We have not been able to determine the
reason why our services have been marked as unhealthy so we are attempting to
simplify the code around the `GazetteerCache` and make failure easier to reason
about.

By adding a custom watchman check that calls `GazetteerCache.get_latest` we
ensure that the task will not be marked as healthy until a gazetteer matcher has
been fully trained.

We increase gunicorn timeouts to 3 minutes to allow for training time now that
the model is no longer being trained on a background thread.